fix robotic moves

+++ Solutions/HackerEarth/Basic Programming/Implementation/Robotic moves/Solution.cs
@@ -14,7 +14,7 @@ namespace CompetitiveProgramming.HackerEarth.Roboticmoves
             int t = int.Parse(Console.ReadLine()!);
             for (int i = 0; i < t; i++)
             {
-                int n = int.Parse(Console.ReadLine()!);
+                long n = long.Parse(Console.ReadLine()!);
                 Console.WriteLine(n * (n + 1));
             }
         }
